+++ include/ftascii.h
@@ -30,6 +30,20 @@ typedef struct player_s
     int         dy;
 }               player_t;
 
+/** @brief  Structure to store the terminal data
+ *  @param  MAX_COL     Maximum number of columns in the terminal
+ *  @param  MAX_ROW     Maximum number of rows in the terminal
+ *  @param  size        Size of the buffer
+ *  @param  buffer      Buffer to store the terminal data
+ *  @param  buffer_copy Copy of the buffer
+ *  @param  pixels      Array of pixels
+ *  @param  frame       Frame number
+ *  @param  clear       Clear the terminal
+ *  @param  delay       Delay between frames
+ *  @param  draw        Draw the terminal
+ *  @param  players     Array of players
+ */
+
 typedef struct term_s
 {
     int             MAX_COL;
@@ -54,7 +68,7 @@ void                init_term(term_t *t);
 void                handleKeyPress(char key, term_t *t);
 void                ft_keyhook(term_t *t);
 
-void                draw(term_t *t);
+void                draw(term_t *t, void (*f)(term_t*));
 void                move_player(term_t *t);
 
 /*     utils       */
@@ -69,7 +83,12 @@ void                fill_pixel(Pixel* pixels, char* color, char* uni, int i);
 
 int                 str_to_hex(char *str);
 void                putpix(Pixel* pixels, char* color, char* uni);
+void                map_pix(term_t *t, int x, int y, char *color, char *uni);
 
 char*               hex_to_unicode(int hex);
+
+// define a list of colors
+extern char*               all_colors[8];
+
 
 #endif
